+++ QMPlusServer/model/socialInsurance/social_order.go
@@ -12,10 +12,10 @@ import (
 )
 
 type OrderReqInfo struct {
-	RemoteAddr string `json:"remoteaddr"`
-	Openid     string `json:"openid"`
-	OrderNo    string `json:"orderno"`
-	TotalFee   int    `json:"totalfee"`
+	RemoteAddr string  `json:"remoteaddr"`
+	Openid     string  `json:"openid"`
+	OrderNo    string  `json:"orderno"`
+	TotalFee   float64 `json:"totalfee"`
 }
 
 type ReqAddOrder struct {
